Fps camera constructor parameters added. Started working on normal mapping. Texture importing fix.
Former-commit-id: 4be3a7cbeeed762ca96e2472900cfcc14bd5a231

+++ PrimeEngine/PrimeEngine-Core/Math/Matrix4x4.cpp
@@ -184,8 +184,8 @@ namespace PrimeEngine { namespace Math {
 	const Matrix4x4 Matrix4x4::Perspective(float fov, float aspectRatio, float zNear, float zFar)
 	{
 		Matrix4x4 result(new float[4][4]{
-			{ (1 / tan(ToRadians(fov) / 2)) / aspectRatio, 0, 0, 0 },
-			{ 0, 1 / tan(ToRadians(fov) / 2), 0, 0 },
+			{ (1 / (float)tan(ToRadians(fov) / 2)) / aspectRatio, 0, 0, 0 },
+			{ 0, 1 / (float)tan(ToRadians(fov) / 2), 0, 0 },
 			{ 0, 0, -zFar / (zFar - zNear), - (zNear * zFar) / (zFar - zNear)},
 			{ 0, 0, -1, 0 }
 		});
